+++ tests/visualization/test_visualization.py
@@ -1,69 +1,94 @@
 
+import pytest
 import numpy as np
 import matplotlib.pyplot as plt
-from dtaianomaly.visualization import plot_time_series_colored_by_score, plot_time_series_anomalies
-import pytest
+from dtaianomaly import visualization
 
-class TestPlotTimeSeriesColoredByScore:
 
-    def test_univariate(self, univariate_time_series):
+@pytest.mark.parametrize('plot_function,additional_args,obligated_y_pred', [
+    (visualization.plot_demarcated_anomalies, {}, False),
+    (visualization.plot_time_series_colored_by_score, {}, False),
+    (visualization.plot_anomaly_scores, {}, True),
+    (visualization.plot_time_series_anomalies, {}, True),
+    (visualization.plot_with_zoom, {'start_zoom': 50, 'end_zoom': 100}, False)
+])
+class TestPlottingFunctions:
+
+    def test_univariate(self, plot_function, additional_args, obligated_y_pred, univariate_time_series):
         y = np.random.choice([0, 1], size=univariate_time_series.shape[0], replace=True)
-        fig = plot_time_series_colored_by_score(univariate_time_series, y)
-        assert len(fig.get_axes()[0].lines) == univariate_time_series.shape[0] - 1
+        if obligated_y_pred:
+            additional_args['y_pred'] = np.random.choice([0, 1], size=univariate_time_series.shape[0], replace=True)
+        plot_function(univariate_time_series, y, **additional_args)
 
-    def test_multivariate(self, multivariate_time_series):
+    def test_multivariate(self, plot_function, additional_args, obligated_y_pred, multivariate_time_series):
         y = np.random.choice([0, 1], size=multivariate_time_series.shape[0], replace=True)
-        fig = plot_time_series_colored_by_score(multivariate_time_series, y)
-        assert len(fig.get_axes()[0].lines) == (multivariate_time_series.shape[0] - 1) * multivariate_time_series.shape[1]
+        if obligated_y_pred:
+            additional_args['y_pred'] = np.random.choice([0, 1], size=multivariate_time_series.shape[0], replace=True)
+        plot_function(multivariate_time_series, y, **additional_args)
 
-    def test_given_axis(self, univariate_time_series):
+    def test_give_time_steps(self, plot_function, additional_args, obligated_y_pred, univariate_time_series):
+        y = np.random.choice([0, 1], size=univariate_time_series.shape[0], replace=True)
+        time_steps = visualization.format_time_steps(None, univariate_time_series.shape[0]) * 2 + 10
+        if obligated_y_pred:
+            additional_args['y_pred'] = np.random.choice([0, 1], size=univariate_time_series.shape[0], replace=True)
+        plot_function(univariate_time_series, y, time_steps=time_steps, **additional_args)
+
+
+@pytest.mark.parametrize('plot_function,additional_args,obligated_y_pred', [
+    (visualization.plot_demarcated_anomalies, {}, False),
+    (visualization.plot_time_series_colored_by_score, {}, False),
+    (visualization.plot_time_series_anomalies, {}, True),
+])
+class TestGivenAxis:
+
+    def test_given_axis(self, plot_function, additional_args, obligated_y_pred, univariate_time_series):
         fig = plt.figure()
         plt.axes()
         [ax] = fig.axes
         y = np.random.choice([0, 1], size=univariate_time_series.shape[0], replace=True)
-        plot_time_series_colored_by_score(univariate_time_series, y, ax=ax)
-        assert len(ax.lines) == univariate_time_series.shape[0] - 1
+        if obligated_y_pred:
+            additional_args['y_pred'] = np.random.choice([0, 1], size=univariate_time_series.shape[0], replace=True)
+        given_fig = plot_function(univariate_time_series, y, ax=ax, **additional_args)
+        assert given_fig == fig
 
-    class TestPlotTimeSeriesAnomalies:
 
-        def test_univariate(self, univariate_time_series):
-            y_true = np.random.choice([0, 1], size=univariate_time_series.shape[0], replace=True)
-            y_pred = np.random.choice([0, 1], size=univariate_time_series.shape[0], replace=True)
-            fig = plot_time_series_anomalies(univariate_time_series, y_true, y_pred)
-            assert len(fig.get_axes()[0].collections) == 3
+@pytest.mark.parametrize('plot_function,additional_args,obligated_y_pred', [
+    (visualization.plot_demarcated_anomalies, {}, False),
+    (visualization.plot_time_series_anomalies, {}, True),
+])
+class TestNonBinaryPrediction:
 
-        def test_multivariate(self, multivariate_time_series):
-            y_true = np.random.choice([0, 1], size=multivariate_time_series.shape[0], replace=True)
-            y_pred = np.random.choice([0, 1], size=multivariate_time_series.shape[0], replace=True)
+    def test_non_binary_y(self, plot_function, additional_args, obligated_y_pred, univariate_time_series):
+        y = np.random.uniform(size=univariate_time_series.shape[0])
+        if obligated_y_pred:
+            additional_args['y_pred'] = np.random.choice([0, 1], size=univariate_time_series.shape[0], replace=True)
+        with pytest.raises(ValueError):
+            plot_function(univariate_time_series, y, **additional_args)
+
+    def test_non_binary_y_pred(self, plot_function, additional_args, obligated_y_pred, univariate_time_series):
+        if obligated_y_pred:
+            y = np.random.choice([0, 1], size=univariate_time_series.shape[0], replace=True)
+            additional_args['y_pred'] = np.random.uniform(size=univariate_time_series.shape[0])
             with pytest.raises(ValueError):
-                plot_time_series_anomalies(multivariate_time_series, y_true, y_pred)
+                plot_function(univariate_time_series, y, **additional_args)
 
-        def test_given_axis(self, univariate_time_series):
-            fig = plt.figure()
-            plt.axes()
-            [ax] = fig.axes
-            y_true = np.random.choice([0, 1], size=univariate_time_series.shape[0], replace=True)
-            y_pred = np.random.choice([0, 1], size=univariate_time_series.shape[0], replace=True)
-            plot_time_series_anomalies(univariate_time_series, y_true, y_pred, ax=ax)
-            assert len(ax.collections) == 3
 
-        def test_correct_scatter_counts(self, univariate_time_series):
-            y_true = np.random.choice([0, 1], size=univariate_time_series.shape[0], replace=True)
-            y_pred = np.random.choice([0, 1], size=univariate_time_series.shape[0], replace=True)
+@pytest.mark.parametrize('plot_function,additional_args,obligated_y_pred', [
+    (visualization.plot_demarcated_anomalies, {}, False),
+    (visualization.plot_time_series_colored_by_score, {}, False),
+    (visualization.plot_time_series_anomalies, {}, True),
+])
+class TestPlotWithZoom:
 
-            TP = (y_true == 1) & (y_pred == 1)
-            FP = (y_true == 0) & (y_pred == 1)
-            FN = (y_true == 1) & (y_pred == 0)
+    def test(self, plot_function, additional_args, obligated_y_pred, univariate_time_series):
+        y = np.random.choice([0, 1], size=univariate_time_series.shape[0], replace=True)
+        if obligated_y_pred:
+            additional_args['y_pred'] = np.random.choice([0, 1], size=univariate_time_series.shape[0], replace=True)
+        visualization.plot_with_zoom(univariate_time_series, y, start_zoom=100, end_zoom=200, method_to_plot=plot_function, **additional_args)
 
-            fig = plot_time_series_anomalies(univariate_time_series, y_true, y_pred)
-            scatter_dots = fig.get_axes()[0].collections
-
-            assert len(scatter_dots[0].get_offsets()) == TP.sum(), "Mismatch in TP count"
-            assert len(scatter_dots[1].get_offsets()) == FP.sum(), "Mismatch in FP count"
-            assert len(scatter_dots[2].get_offsets()) == FN.sum(), "Mismatch in FN count"
-
-        def test_non_binary_y_pred(self, univariate_time_series):
-            y_true = np.random.choice([0, 1], size=univariate_time_series.shape[0], replace=True)
-            y_pred = np.random.uniform(0, 1, size=univariate_time_series.shape[0])
-            with pytest.raises(ValueError, match="The predicted anomaly scores must be binary."):
-                plot_time_series_anomalies(univariate_time_series, y_true, y_pred)
+    def test_given_time_steps(self, plot_function, additional_args, obligated_y_pred, univariate_time_series):
+        y = np.random.choice([0, 1], size=univariate_time_series.shape[0], replace=True)
+        time_steps = visualization.format_time_steps(None, univariate_time_series.shape[0]) * 2 + 10
+        if obligated_y_pred:
+            additional_args['y_pred'] = np.random.choice([0, 1], size=univariate_time_series.shape[0], replace=True)
+        visualization.plot_with_zoom(univariate_time_series, y, start_zoom=100, end_zoom=200, time_steps=time_steps, method_to_plot=plot_function, **additional_args)